SASS-11611 refactor SessionDataHelper and TaxYearHelper
