+++ app/utils/TaxYearHelper.scala
@@ -39,12 +39,4 @@ trait TaxYearHelper extends SessionDataHelper {
 
   val taxYear: Int = if (dateNow.isAfter(taxYearCutoffDate)) LocalDate.now().getYear + 1 else LocalDate.now().getYear
 
-  def retrieveTaxYearList(implicit user: User[_]): Seq[Int] = {
-    user.session.get(SessionValues.VALID_TAX_YEARS).getOrElse("").split(",").toSeq.map(_.toInt)
-  }
-
-  def firstClientTaxYear(implicit user: User[_]): Int = retrieveTaxYearList.head
-  def latestClientTaxYear(implicit user: User[_]): Int = retrieveTaxYearList.last
-
-  def singleValidTaxYear(implicit user: User[_]): Boolean = firstClientTaxYear == latestClientTaxYear
 }
